Huge Update
- Removed all indexes outside of the used one (Index2). Renamed to Index and moved into components folder.

- Updated react-router-dom to 6.2.1 REQUIRES A RE-INSTALL

- Added routing to App (and converted to functional component)

- scrollSpy has some janky fixes but is difficult to convert to a functional component
