+++ Landing/package.json
@@ -15,7 +15,7 @@
     "react-image-lightbox": "^5.1.1",
     "react-modal-video": "^1.2.3",
     "react-parallax": "^3.0.3",
-    "react-router-dom": "^5.1.2",
+    "react-router-dom": "^6.2.1",
     "react-scripts": "3.4.1",
     "react-slick": "^0.26.1",
     "react-sticky-header": "^0.2.0",
