+++ Landing/src/components/Inquire/Inquire.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+
+
+const Inquire = () => {
+    
+  return(
+    <div style={{marginTop: '4rem', width: '80%', margin: '8rem auto', backgroundColor: 'grey', padding: '2rem'}}>
+      <h1>Hey its an inquiry form.</h1>
+    </div>
+  )
+};
+
+export default Inquire;
