+++ Landing/src/components/Navbar/scrollSpy.js
@@ -1,4 +1,5 @@
 import React, { Component } from 'react';
+import { useLocation } from 'react-router-dom';
 
 class ScrollspyNav extends Component {
     constructor(props) {
@@ -85,22 +86,26 @@ class ScrollspyNav extends Component {
     scrollSection = () => {
             let scrollSectionOffsetTop;
             this.scrollTargetIds.forEach((sectionID, index) => {
-                scrollSectionOffsetTop = document.getElementById(sectionID).offsetTop - (this.headerBackground ? document.querySelector("div[data-nav='list']").scrollHeight : 0);
-
-                if (window.pageYOffset >= scrollSectionOffsetTop && window.pageYOffset < scrollSectionOffsetTop + document.getElementById(sectionID).scrollHeight) {
-                    this.getNavLinkElement(sectionID).classList.add(this.activeNavClass);
-                    this.getNavLinkElement(sectionID).parentNode.classList.add(this.activeNavClass);
-                    this.clearOtherNavLinkActiveStyle(sectionID)
-                } else {
-                    this.getNavLinkElement(sectionID).classList.remove(this.activeNavClass);
-                    this.getNavLinkElement(sectionID).parentNode.classList.remove(this.activeNavClass);
-                }
-
-                if (window.innerHeight + window.pageYOffset >= document.body.scrollHeight && index === this.scrollTargetIds.length - 1) {
-                    this.getNavLinkElement(sectionID).classList.add(this.activeNavClass);
-                    this.getNavLinkElement(sectionID).parentNode.classList.add(this.activeNavClass);
-                    this.clearOtherNavLinkActiveStyle(sectionID);
-                }
+                if(document.getElementById(sectionID) === null){
+                    scrollSectionOffsetTop = 0
+                    return;
+                } else{
+                    scrollSectionOffsetTop = document.getElementById(sectionID).offsetTop - (this.headerBackground ? document.querySelector("div[data-nav='list']").scrollHeight : 0);
+                    if (window.pageYOffset >= scrollSectionOffsetTop && window.pageYOffset < scrollSectionOffsetTop + document.getElementById(sectionID).scrollHeight) {
+                        this.getNavLinkElement(sectionID).classList.add(this.activeNavClass);
+                        this.getNavLinkElement(sectionID).parentNode.classList.add(this.activeNavClass);
+                        this.clearOtherNavLinkActiveStyle(sectionID)
+                    } else {
+                        this.getNavLinkElement(sectionID).classList.remove(this.activeNavClass);
+                        this.getNavLinkElement(sectionID).parentNode.classList.remove(this.activeNavClass);
+                    }
+    
+                    if (window.innerHeight + window.pageYOffset >= document.body.scrollHeight && index === this.scrollTargetIds.length - 1) {
+                        this.getNavLinkElement(sectionID).classList.add(this.activeNavClass);
+                        this.getNavLinkElement(sectionID).parentNode.classList.add(this.activeNavClass);
+                        this.clearOtherNavLinkActiveStyle(sectionID);
+                    }
+                }               
             });
     }
     clearOtherNavLinkActiveStyle(excludeSectionID) {
